move NotificationBuilder to ServiceNotification and require it as a constructor argument from TorServiceController.Builder
