+++ sampleapp/src/main/java/io/matthewnelson/sampleapp/App.kt
@@ -17,7 +17,11 @@
 package io.matthewnelson.sampleapp
 
 import android.app.Application
+import android.content.Context
+import io.matthewnelson.topl_core_base.TorConfigFiles
 import io.matthewnelson.topl_service.TorServiceController
+import io.matthewnelson.topl_service.notification.ServiceNotification
+import java.io.File
 
 /**
  * @suppress
@@ -26,23 +30,17 @@ class App: Application() {
 
     override fun onCreate() {
         super.onCreate()
-        setupTorServices()
+        setupTorServices(this, TorConfigFiles.createConfig(this))
     }
 
-    private fun setupTorServices() {
-        TorServiceController.Builder(
-            application = this,
-            buildConfigVersionCode = BuildConfig.VERSION_CODE,
-            torSettings = MyTorSettings(),
-            geoipAssetPath = "common/geoip",
-            geoip6AssetPath = "common/geoip6"
+    private fun generateTorServiceNotificationBuilder(): ServiceNotification.Builder {
+//  private fun generateTorServiceNotificationBuilder(): ServiceNotification.Builder {
+        return ServiceNotification.Builder(
+            channelName = "TorService Channel",
+            channelDescription = "Tor Channel",
+            channelID = "My Sample Application",
+            notificationID = 615
         )
-
-            .customizeNotification(
-                channelName = "TorService Channel",
-                channelDescription = "Tor Channel",
-                channelID = "My Sample Application",
-                notificationID = 615)
             .setActivityToBeOpenedOnTap(
                 clazz = MainActivity::class.java,
                 intentExtrasKey = null,
@@ -50,13 +48,69 @@ class App: Application() {
                 intentRequestCode = null
             )
             .setCustomColor(colorRes = R.color.primaryColor)
-            .enableTorRestartButton()
-            .enableTorStopButton()
-            .showNotification(false)
-            .applyNotificationSettings()
+            .enableTorRestartButton(enable = true)
+            .enableTorStopButton(enable = true)
+            .showNotification(show = true)
+//  }
+    }
 
+    private fun setupTorServices(application: Application, torConfigFiles: TorConfigFiles) {
+//  private fun setupTorServices(application: Application, torConfigFiles: TorConfigFiles ) {
+        TorServiceController.Builder(
+            application = application,
+            torServiceNotificationBuilder = generateTorServiceNotificationBuilder(),
+            buildConfigVersionCode = BuildConfig.VERSION_CODE,
+
+            // Can instantiate directly here then access it from
+            // TorServiceController.Companion.getTorSettings() and cast what's returned
+            // as MyTorSettings
+            torSettings = MyTorSettings(),
+
+            // These should live somewhere in your module's assets directory,
+            // ex: my-project/my-application-module/src/main/assets/common/geoip
+            // ex: my-project/my-application-module/src/main/assets/common/geoip6
+            geoipAssetPath = "common/geoip",
+            geoip6AssetPath = "common/geoip6"
+        )
             .setBuildConfigDebug(buildConfigDebug = BuildConfig.DEBUG)
+
+            // Can instantiate directly here then access it from
+            // TorServiceController.Companion?.appEventBroadcaster and cast what's returned
+            // as MyEventBroadcaster
             .setEventBroadcaster(eventBroadcaster = MyEventBroadcaster())
+
+            // Only needed if you wish to customize the directories/files used by Tor if
+            // the defaults aren't to your liking.
+            .useCustomTorConfigFiles(torConfigFiles = torConfigFiles)
+
             .build()
+//  }
+    }
+
+    fun customTorConfigFilesSetup(context: Context): TorConfigFiles {
+//  fun customTorConfigFilesSetup(context: Context): TorConfigFiles {
+
+        // This is modifying the directory hierarchy from TorService's
+        // default setup. For example, if you are using binaries for Tor that
+        // are named differently that that expressed in TorConfigFiles.createConfig()
+
+        // Post Android API 28 requires that executable files be contained in your
+        // application's data/app directory, as they can no longer execute from data/data.
+        val installDir = File(context.applicationInfo.nativeLibraryDir)
+
+        // Will create a directory within your application's data/data dir
+        val configDir = context.getDir("torservice", Context.MODE_PRIVATE)
+
+        val builder = TorConfigFiles.Builder(installDir, configDir)
+
+        // Customize the tor executable file name. Requires that the executable file
+        // be in your module's src/main/jniLibs directory. If you are getting your
+        // executable files via a dependency be sure to consult that Library's documentation.
+        builder.torExecutable(File(installDir, "libtor.so"))
+
+        // customize further via the builder methods...
+
+        return builder.build()
+//  }
     }
 }
