+++ topl-core-base/src/main/java/io/matthewnelson/topl_core_base/TorConfigFiles.kt
@@ -199,7 +199,7 @@ class TorConfigFiles private constructor(
      *
      * @param [installDir] directory where the tor binaries are installed.
      * @param [configDir] directory where the filesystem will be setup for tor.
-     * @sample [io.matthewnelson.sampleapp.samplecode.SampleCode.customTorConfigFilesSetup]
+     * @sample [io.matthewnelson.sampleapp.App.customTorConfigFilesSetup]
      */
     class Builder(private val installDir: File, private val configDir: File) {
