+++ topl-service/build.gradle
@@ -44,7 +44,7 @@ dokka {
         includeNonPublic = false
         skipEmptyPackages = true
         samples = [
-                "$rootDir/sampleapp/src/main/java/io/matthewnelson/sampleapp/samplecode/SampleCode.kt".toString()
+                "$rootDir/sampleapp/src/main/java/io/matthewnelson/sampleapp/App.kt".toString()
         ]
         sourceLink {
             url = "https://github.com/05nelsonm/TorOnionProxyLibrary-Android/blob/master/"
