+++ topl-service/src/main/java/io/matthewnelson/topl_service/TorServiceController.kt
@@ -59,28 +59,28 @@ class TorServiceController private constructor(): ServiceConsts() {
      * [io.matthewnelson.topl_service.onionproxy.ServiceTorSettings] and
      * [TorService.initTOPLCore].
      *
-     * @param [application] [Application], for obtaining context.
+     * @param [application] [Application], for obtaining context
+     * @param [torServiceNotificationBuilder] The [ServiceNotification.Builder] for
+     *   customizing [TorService]'s notification
      * @param [buildConfigVersionCode] send [BuildConfig.VERSION_CODE]. Mitigates copying of geoip
-     *   files to app updates only.
-     * @param [torSettings] [TorSettings] used to create your torrc file on start of Tor.
+     *   files to app updates only
+     * @param [torSettings] [TorSettings] used to create your torrc file on start of Tor
      * @param [geoipAssetPath] The path to where you have your geoip file located (ex: in
-     *   assets/common directory, send this variable "common/geoip").
+     *   assets/common directory, send this variable "common/geoip")
      * @param [geoip6AssetPath] The path to where you have your geoip6 file located (ex: in
-     *   assets/common directory, send this variable "common/geoip6").
+     *   assets/common directory, send this variable "common/geoip6")
      *
-     * @sample [io.matthewnelson.sampleapp.samplecode.SampleCode.setupTorServices]
+     * @sample [io.matthewnelson.sampleapp.App.generateTorServiceNotificationBuilder]
+     * @sample [io.matthewnelson.sampleapp.App.setupTorServices]
      * */
     class Builder(
         private val application: Application,
+        private val torServiceNotificationBuilder: ServiceNotification.Builder,
         private val buildConfigVersionCode: Int,
-        torSettings: TorSettings,
+        private val torSettings: TorSettings,
         private val geoipAssetPath: String,
         private val geoip6AssetPath: String
     ) {
-
-        init {
-            Companion.torSettings = torSettings
-        }
 
         private lateinit var torConfigFiles: TorConfigFiles
 
@@ -96,11 +96,11 @@ class TorServiceController private constructor(): ServiceConsts() {
          * will still get the same messages sent to your [EventBroadcaster] if you set it
          * via [Builder.setEventBroadcaster].
          *
-         * TODO: Provide a link to gh-pages that discusses logging and how it work, it's pretty
-         *  complex with everything that is going on.
-         *
          * @param [buildConfigDebug] Send [BuildConfig.DEBUG]
          * @see [io.matthewnelson.topl_core.broadcaster.BroadcastLogger]
+         *
+         * TODO: Provide a link to gh-pages that discusses logging and how it work, it's pretty
+         *  complex with everything that is going on.
          * */
         fun setBuildConfigDebug(buildConfigDebug: Boolean): Builder {
             this.buildConfigDebug = buildConfigDebug
@@ -130,7 +130,7 @@ class TorServiceController private constructor(): ServiceConsts() {
          * to operate with.
          *
          * @return [Builder]
-         * @sample [io.matthewnelson.sampleapp.samplecode.SampleCode.customTorConfigFilesSetup]
+         * @sample [io.matthewnelson.sampleapp.App.customTorConfigFilesSetup]
          * @see [Builder.build]
          * */
         fun useCustomTorConfigFiles(torConfigFiles: TorConfigFiles): Builder {
@@ -139,257 +139,16 @@ class TorServiceController private constructor(): ServiceConsts() {
         }
 
         /**
-         * Customize the service notification to your application.
-         *
-         * See [Builder] for code samples.
-         *
-         * @param [channelName] Your notification channel's name (cannot be empty).
-         * @param [channelID] Your notification channel's ID (cannot be empty).
-         * @param [channelDescription] Your notification channel's description (cannot be empty).
-         * @param [notificationID] Your foreground notification's ID.
-         * @return [NotificationBuilder] To obtain methods specific to notification customization.
-         * @throws [IllegalArgumentException] If String fields are empty.
-         * */
-        @Throws(IllegalArgumentException::class)
-        fun customizeNotification(
-            channelName: String,
-            channelID: String,
-            channelDescription: String,
-            notificationID: Short
-        ): NotificationBuilder {
-            require(
-                channelName.isNotEmpty() &&
-                channelID.isNotEmpty() &&
-                channelDescription.isNotEmpty()
-            ) { "channelName, channelID, & channelDescription must not be empty." }
-
-            return NotificationBuilder(
-                this,
-                channelName,
-                channelID,
-                channelDescription,
-                notificationID.toInt()
-            )
-        }
-
-        /**
-         * Where you get to customize how your foreground notification will look/function.
-         * Calling [customizeNotification] will return this class to you which provides methods
-         * specific to customization of notifications. Call [applyNotificationSettings] when done
-         * to return to [Builder] to continue with it's methods for customization.
-         *
-         * See [Builder] for code samples.
-         *
-         * @param [builder] [Builder] To return to it when calling [applyNotificationSettings]
-         * @param [channelName] Your notification channel's name.
-         * @param [channelID] Your notification channel's ID.
-         * @param [channelDescription] Your notification channel's description.
-         * @param [notificationID] Your foreground notification's ID.
-         * */
-        class NotificationBuilder(
-            private val builder: Builder,
-            channelName: String,
-            channelID: String,
-            channelDescription: String,
-            notificationID: Int
-        ) {
-
-            private val serviceNotification =
-                ServiceNotification(
-                    channelName,
-                    channelID,
-                    channelDescription,
-                    notificationID
-                )
-
-            /**
-             * For when your user taps the TorService notification.
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [clazz] The Activity to be opened when tapped.
-             * @param [intentExtrasKey]? The key for if you with to add extras in the PendingIntent.
-             * @param [intentExtras]? The extras that will be sent in the PendingIntent.
-             * @param [intentRequestCode]? The request code - Defaults to 0 if not set.
-             *
-             * TODO: Check if this conflicts with multi-activity apps such that if the user is
-             *  on a certain activity, and clicks the notification it will bring them to the
-             *  declared activity instead of back to the activity they were on.
-             *  + Include an optional Bundle? to be set for creating the pending intent.
-             *  + Think about overriding and providing another option to rotate the ContentIntent
-             *  to open up/resume current activity?
-             * */
-            fun setActivityToBeOpenedOnTap(
-                clazz: Class<*>,
-                intentExtrasKey: String?,
-                intentExtras: String?,
-                intentRequestCode: Int?
-            ): NotificationBuilder {
-                serviceNotification.activityWhenTapped = clazz
-                serviceNotification.activityIntentKey = intentExtrasKey
-                serviceNotification.activityIntentExtras = intentExtras
-                intentRequestCode?.let { serviceNotification.activityIntentRequestCode = it }
-                return this
-            }
-
-            /**
-             * Defaults to Orbot/TorBrowser's icon [R.drawable.tor_stat_network_enabled].
-             *
-             * The small icon you wish to display when Tor's network state is
-             * [io.matthewnelson.topl_core_base.BaseConsts.TorNetworkState.ENABLED].
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [drawableRes] Drawable resource id.
-             * @return [NotificationBuilder]
-             * */
-            fun setImageTorNetworkingEnabled(@DrawableRes drawableRes: Int): NotificationBuilder {
-                serviceNotification.imageNetworkEnabled = drawableRes
-                return this
-            }
-
-            /**
-             * Defaults to Orbot/TorBrowser's icon [R.drawable.tor_stat_network_disabled].
-             *
-             * The small icon you wish to display when Tor's network state is
-             * [io.matthewnelson.topl_core_base.BaseConsts.TorNetworkState.DISABLED].
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [drawableRes] Drawable resource id.
-             * @return [NotificationBuilder]
-             * */
-            fun setImageTorNetworkingDisabled(@DrawableRes drawableRes: Int): NotificationBuilder {
-                serviceNotification.imageNetworkDisabled = drawableRes
-                return this
-            }
-
-            /**
-             * Defaults to Orbot/TorBrowser's icon [R.drawable.tor_stat_network_dataxfer].
-             *
-             * The small icon you wish to display when bandwidth is being used.
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [drawableRes] Drawable resource id.
-             * @return [NotificationBuilder]
-             * */
-            fun setImageTorDataTransfer(@DrawableRes drawableRes: Int): NotificationBuilder {
-                serviceNotification.imageDataTransfer = drawableRes
-                return this
-            }
-
-            /**
-             * Defaults to Orbot/TorBrowser's icon [R.drawable.tor_stat_notifyerr].
-             *
-             * The small icon you wish to display when Tor is having problems.
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [drawableRes] Drawable resource id.
-             * @return [NotificationBuilder]
-             * */
-            fun setImageTorErrors(@DrawableRes drawableRes: Int): NotificationBuilder {
-                serviceNotification.imageError = drawableRes
-                return this
-            }
-
-            /**
-             * Defaults to [R.color.tor_service_white]
-             *
-             * The color you wish to display when Tor's network state is
-             * [io.matthewnelson.topl_core_base.BaseConsts.TorNetworkState.ENABLED].
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [colorRes] Color resource id.
-             * @return [NotificationBuilder]
-             * */
-            fun setCustomColor(@ColorRes colorRes: Int): NotificationBuilder {
-                serviceNotification.colorWhenConnected = colorRes
-                return this
-            }
-
-            /**
-             * Defaults to NotificationVisibility.VISIBILITY_SECRET
-             *
-             * The visibility of your notification on the user's lock screen.
-             *
-             * See [Builder] for code samples.
-             *
-             * @return [NotificationBuilder]
-             * */
-            fun setVisibility(@NotificationVisibility visibility: Int): NotificationBuilder {
-                if (visibility in -1..1)
-                    serviceNotification.visibility = visibility
-                return this
-            }
-
-            /**
-             * Disabled by Default
-             *
-             * Enable on the notification the ability to *restart* Tor.
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [enable] Boolean, automatically set to true but provides cleaner option
-             *   for implementor to query SharedPreferences for user's settings (if desired).
-             * @return [NotificationBuilder]
-             * */
-            fun enableTorRestartButton(enable: Boolean = true): NotificationBuilder {
-                serviceNotification.enableRestartButton = enable
-                return this
-            }
-
-            /**
-             * Disabled by Default
-             *
-             * Enable on the notification the ability to *stop* Tor.
-             *
-             * See [Builder] for code samples.
-             *
-             * @param [enable] Boolean, automatically set to true but provides cleaner option
-             *   for implementor to query SharedPreferences for user's settings (if desired).
-             * @return [NotificationBuilder]
-             * */
-            fun enableTorStopButton(enable: Boolean = true): NotificationBuilder {
-                serviceNotification.enableStopButton = enable
-                return this
-            }
-
-            /**
-             * Enabled by Default.
-             *
-             * Setting it to false will only show a notification when your task is removed
-             * in order to properly shutdown the Service.
-             * */
-            fun showNotification(show: Boolean = false): NotificationBuilder {
-                serviceNotification.showNotification = show
-                return this
-            }
-
-            /**
-             * Initialize settings.
-             *
-             * See [Builder] for code samples.
-             *
-             * @return [Builder]
-             * */
-            fun applyNotificationSettings(): Builder {
-                ServiceNotification.initialize(serviceNotification)
-                return builder
-            }
-
-        }
-
-        /**
-         * Initializes [TorService] setup and enables the ability to call methods in the
-         * [Companion] object.
+         * Initializes [TorService] setup and enables the ability to call methods from the
+         * [Companion] object w/o throwing exceptions.
          *
          * See [Builder] for code samples.
          * */
         fun build() {
 
+            torServiceNotificationBuilder.build()
+
+            Companion.torSettings = torSettings
             Companion.torConfigFiles =
                 if (::torConfigFiles.isInitialized)
                     torConfigFiles
